backend/auto-server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
mysql-connector-python==8.2.0
openai==1.3.0
pybit==5.7.0
cryptography==41.0.7
python-dotenv==1.0.0
pydantic==2.5.0
schedule==1.2.0
asyncio==3.4.3